refactor: update api
