+++ src/api.ts
@@ -1,5 +1,6 @@
-export * as TypeScriptCodeGenerator from "./internal/CodeGenerator";
+export * as TsGenerator from "./internal/CodeGenerator";
 export * as Converter from "./internal/Converter";
 export { fileSystem } from "./internal/FileSystem";
 export * as ResolveReference from "./internal/ResolveReference";
 export * as Validator from "./internal/Validator";
+export * as Factory from "./internal/CodeGenerator/factory";
