+++ src/code-templates/api-client/ApiClientClass/MethodBody/__tests__/PathParameter-test.ts
@@ -2,7 +2,7 @@ import { EOL } from "os";
 
 import ts from "typescript";
 
-import { Factory } from "../../../../../factory";
+import { Factory } from "../../../../../api";
 import type { CodeGenerator } from "../../../../../types";
 import * as Utils from "../../../utils";
 import * as PathParameter from "../PathParameter";
