+++ src/code-templates/api-client/ApiClientClass/MethodBody/index.ts
@@ -1,6 +1,6 @@
 import ts from "typescript";
 
-import type { Factory } from "../../../../factory";
+import type { Factory } from "../../../../api";
 import type { CodeGenerator } from "../../../../types";
 import * as Utils from "../../utils";
 import * as CallRequest from "./CallRequest";
