--- src/factory.ts
@@ -1,1 +1,0 @@
-export * as Factory from "./internal/CodeGenerator/factory";
